[CSE-34] SUM이 null이 되는 문제 해결

+++ modules/draw/draw-api/src/main/java/com/eodi/bium/draw/view/DrawPointView.java
@@ -2,7 +2,7 @@ package com.eodi.bium.draw.view;
 
 public record DrawPointView(
     String memberId,
-    Long point
+    Integer point
 ) {
 
 }
